requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.8
ikea_api==2.3.0
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
